+++ src/Helpers/GatewayHelper.php
@@ -5,6 +5,7 @@ namespace Buckaroo\Shopware6\Helpers;
 use Buckaroo\Shopware6\PaymentMethods\Ideal;
 use Buckaroo\Shopware6\PaymentMethods\IdealProcessing;
 use Buckaroo\Shopware6\PaymentMethods\Bancontact;
+use Buckaroo\Shopware6\PaymentMethods\Creditcard;
 use Buckaroo\Shopware6\PaymentMethods\Creditcards;
 use Buckaroo\Shopware6\PaymentMethods\AfterPay;
 use Buckaroo\Shopware6\PaymentMethods\Sofort;
@@ -24,6 +25,7 @@ class GatewayHelper
         Ideal::class,
         IdealProcessing::class,
         Bancontact::class,
+        Creditcard::class,
         Creditcards::class,
         AfterPay::class,
         Sofort::class,