+++ src/Resources/app/storefront/src/main.js
@@ -1,0 +1,4 @@
+import BuckarooPaymentCreditcards from './creditcards/buckaroo-payment.creditcards';
+
+const PluginManager = window.PluginManager;
+PluginManager.register('BuckarooPaymentCreditcards', BuckarooPaymentCreditcards);